Fix: PR 기반으로 수정사항들
- EOF
- changeState -> delete
- CRUD -> QueryService + CommandService 로 분리
